+++ lib/redgreenblue.rb
@@ -8,6 +8,7 @@ require 'redgreenblue/base'
 require 'redgreenblue/24bit'
 require 'redgreenblue/48bit'
 require 'redgreenblue/hex'
+require 'redgreenblue/int'
 
 require 'redgreenblue/hsl_hsv'
 require 'redgreenblue/cie'