+++ History.md
@@ -14,6 +14,8 @@
 
     * Added `#to_i` method.
 
+    * Added `::assemble` factory method.
+    
     * Added `::at` factory method.
     
     * Added `::yellow`, `::cyan`, and `::magenta` factory methods.