+++ History.md
@@ -1,6 +1,6 @@
 # History
 
-## Changes
+## 0.9.0 (2 March 2020)
 
  * __General:__